+++ Runtime/SceneManager/SceneManager.cs
@@ -1,8 +1,5 @@
 using System;
-using System.Collections;
-using System.Threading.Tasks;
 using UnityEngine;
-using ActionCode.AwaitableCoroutines;
 using SceneMode = UnityEngine.SceneManagement.LoadSceneMode;
 using UnitySceneManager = UnityEngine.SceneManagement.SceneManager;
 
@@ -83,8 +80,8 @@ namespace ActionCode.SceneManagement
         /// </summary>
         /// <param name="scene"><inheritdoc cref="LoadScene(string, SceneTransition)" path="/param[@name='scene']"/></param>
         /// <param name="transition"><inheritdoc cref="LoadScene(string, SceneTransition)" path="/param[@name='transition']"/></param>
-        /// <returns>An asynchronously Task.</returns>
-        public static async Task LoadSceneAsync(string scene, SceneTransition transition = null) =>
+        /// <returns>An Awaitable Task.</returns>
+        public static async Awaitable LoadSceneAsync(string scene, SceneTransition transition = null) =>
             await LoadSceneAsync(new Scene(scene), transition);
 
         /// <summary>
@@ -93,10 +90,7 @@ namespace ActionCode.SceneManagement
         /// <param name="scene"><inheritdoc cref="LoadScene(Scene, SceneTransition)" path="/param[@name='scene']"/></param>
         /// <param name="transition"><inheritdoc cref="LoadScene(string, SceneTransition)" path="/param[@name='transition']"/></param>
         /// <returns><inheritdoc cref="LoadSceneAsync(string, SceneTransition)"/></returns>
-        public static async Task LoadSceneAsync(Scene scene, SceneTransition transition = null) =>
-            await AwaitableCoroutine.Run(LoadSceneCoroutine(scene, transition));
-
-        private static IEnumerator LoadSceneCoroutine(Scene scene, SceneTransition transition)
+        public static async Awaitable LoadSceneAsync(Scene scene, SceneTransition transition = null)
         {
             Time.timeScale = 1f;
 
@@ -112,7 +106,7 @@ namespace ActionCode.SceneManagement
 
             var hasLoadingScene = transition.HasLoadingScene();
 
-            yield return transition.ScreenFader?.FadeOut();
+            await transition.ScreenFader?.FadeOutAsync();
             IProgress<float> progress = new Progress<float>(ReportProgress);
 
             if (hasLoadingScene)
@@ -124,16 +118,16 @@ namespace ActionCode.SceneManagement
                 if (loadingSceneOperation == null)
                 {
                     LoadingScene = null;
-                    yield break;
+                    return;
                 }
 
-                yield return loadingSceneOperation;
+                await loadingSceneOperation;
 
                 progress.Report(0F);
-                yield return transition.ScreenFader?.FadeIn();
+                await transition.ScreenFader?.FadeInAsync();
             }
 
-            yield return new WaitForSeconds(transition.TimeBeforeLoading);
+            await Awaitable.WaitForSecondsAsync(transition.TimeBeforeLoading);
 
             var loadingOperation = UnitySceneManager.LoadSceneAsync(scene.path);
 
@@ -141,31 +135,36 @@ namespace ActionCode.SceneManagement
             if (loadingOperation == null)
             {
                 LoadingScene = null;
-                yield break;
+                return;
             }
 
             // Prevent to automatically unload the LoadingScene if any.
             loadingOperation.allowSceneActivation = false;
 
-            yield return loadingOperation.WaitUntilActivationProgress(progress);
+            await loadingOperation.WaitUntilActivationProgress(progress);
 
             progress.Report(1F);
-            yield return new WaitForSeconds(transition.TimeAfterLoading);
-            yield return new WaitUntil(() => !IsLoadingLocked);
 
-            if (hasLoadingScene) yield return transition.ScreenFader?.FadeOut();
+            await Awaitable.WaitForSecondsAsync(transition.TimeAfterLoading);
+            await WaitUntilAsync(() => !IsLoadingLocked);
+
+            if (hasLoadingScene) await transition.ScreenFader?.FadeOutAsync();
 
             // Automatically unload the LoadingScene if any.
             loadingOperation.allowSceneActivation = true;
 
-            yield return new WaitUntil(() => loadingOperation.isDone);
-
-            yield return transition.ScreenFader?.FadeIn();
+            await WaitUntilAsync(() => loadingOperation.isDone);
+            await transition.ScreenFader?.FadeInAsync();
 
             LoadingScene = null;
             OnLoadingFinished?.Invoke();
         }
 
         private static void ReportProgress(float progress) => OnProgressChanged?.Invoke(progress * 100F);
+
+        private static async Awaitable WaitUntilAsync(Func<bool> condition)
+        {
+            while (!condition()) await Awaitable.NextFrameAsync();
+        }
     }
 }